apps/langchain/requirements.txt
langchain
langchain-cli
langchain-experimental
langsmith
fastapi
uvicorn
debugpy
openai
python-dotenv
google-search-results